requirements.txt
fastapi
crewai
crewai[tools]
langchain
langchain_core
langchain-community
uvicorn[standard]
python-dotenv
tqdm
tiktoken
langchain-openai 
wikipedia 
tavily-python 
langgraph
google-search-results